+++ sdk/storage/azure-storage-queue/src/test/java/com/azure/storage/queue/QueueAsyncApiTests.java
@@ -807,6 +807,23 @@ public class QueueAsyncApiTests extends QueueTestBase {
     }
 
     @Test
+    public void updateMessageWithBase64Client() {
+        String updateMsg = "Updated test message";
+        QueueAsyncClient encodingQueueClient = getBase64Client();
+        encodingQueueClient.create().block();
+        encodingQueueClient.sendMessage("test message before update").block();
+
+        QueueMessageItem dequeueMsg = encodingQueueClient.receiveMessage().block();
+
+        assertAsyncResponseStatusCode(encodingQueueClient.updateMessageWithResponse(dequeueMsg.getMessageId(),
+            dequeueMsg.getPopReceipt(), updateMsg, Duration.ofSeconds(1)), 204);
+
+        StepVerifier.create(encodingQueueClient.peekMessage().delaySubscription(getMessageUpdateDelay(2000)))
+            .assertNext(peekedMessageItem -> assertEquals(updateMsg, peekedMessageItem.getMessageText()))
+            .verifyComplete();
+    }
+
+    @Test
     public void getQueueName() {
         assertEquals(queueName, queueAsyncClient.getQueueName());
     }
@@ -833,6 +850,11 @@ public class QueueAsyncApiTests extends QueueTestBase {
 
         StepVerifier.create(queueAsyncClient.getPropertiesWithResponse()).assertNext(queuePropertiesResponse ->
             assertEquals("2017-11-09", queuePropertiesResponse.getHeaders().getValue("x-ms-version"))).verifyComplete();
+    }
+
+    private QueueAsyncClient getBase64Client() {
+        return queueServiceBuilderHelper().messageEncoding(QueueMessageEncoding.BASE64)
+            .buildAsyncClient().getQueueAsyncClient(queueName);
     }
 
     @Test